requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-rest-framework-simplejwt==5.0.0
mysqlclient==2.1.0
drf-yasg==1.20.0


